`Discord`: fix `sendMessage` implementation
This commit adjusts the `Discord.sendMessage` Temporal activity implementation
to accept a `:`-delimited string as the `guildChannel`, which is then
split to get the correct `channelId`. This cached channel is correctly
typed as a `TextChannel`, and is sent the provided message if the channel
name is `general`.

Default channel name is `general`. This will be configurable eventually.