Add `Temporal.io` activities/worker implementation (#11)
* Add required `@temporalio` packages

* Implement Temporal worker, register activities

This commit adds the Temporal `Worker` initialization functionality with
appropriate typing for the registered `activities`. Also, the `Platform`
interface and subsequent implementations are modified to include the
handlers for the `Worker` activities (currently `sendMessage`).

* Temporal: add `host` config parameter

Add a `host` parameter to the Temporal config object that is used during
`Worker` initialization to connect to a specific Temporal server.

Defaults to `127.0.0.1`

* `Discord`: fix `sendMessage` implementation

This commit adjusts the `Discord.sendMessage` Temporal activity implementation
to accept a `:`-delimited string as the `guildChannel`, which is then
split to get the correct `channelId`. This cached channel is correctly
typed as a `TextChannel`, and is sent the provided message if the channel
name is `general`.

Default channel name is `general`. This will be configurable eventually.

* `config.ts`: export `ParsedConfig` type

* `Discord`: send message if `instanceof TextChannel`

In this commit, we remove the requirement that the `sendMessage` destination
channel is named "general". Rather, we assume that the admin has configured
provided a `guildId`/`channelId` that is appropriate for the bot's assigned
permissions.

* Only initialize Temporal `Worker` if fully configured

This commit adds a conditional to ensure that all required Temporal config
directives are populated in the `.env` file. If any are `undefined` or
empty strings, the Temporal `Worker` initialization will be skipped.

* Add `temporal/workflows.ts` for `Worker` init

* `Telegra`: Disable link previews in messages

* Move Temporal activities to `temporal/index.ts`

This commit moves the imported `temporal.ts` file to the `lib/temporal`
directory and is renamed to `index.ts` to maintain import paths. It also
exports the `LotusBotActivities` type directly, rather than define/export
a redundant `interface`.

* Convert `LotusBot.temporal.*` to arrow functions

This commit should have been included previously. It converts the method
definitions within the `LotusBot.temporal` object to be arrow functions.
This is necessary for the called function to properly reference `this`
as the instantiated `LotusBot` class, e.g. for calling `sendMessage` on
instantiated platform clients.

* `Temporal`: Add skeleton `sendLotus` activity