Implement bot wallet (#21)
With many of the new features being added across Lotusia, `lotus-bot` is
going to need its own wallet. This branch will check if the bot instance
has its own wallet on start-up, and will create one if it does not. The
wallet will be loaded with all other wallets on subsequent runtimes.